moved date converter to utils
